+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2026-01-14
+
+### Fixed
+
+- [#54](https://github.com/n15g/badger/issues/54) - Links in badge tooltip are broken
+
 ## [2.1.0] - 2026-01-14
 
 ### Added